+++ cognite/client/_api/data_modeling/instances.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 
 import json
-from typing import TYPE_CHECKING, Any, Iterator, List, Literal, Sequence, Union, cast, overload
+from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Literal, Sequence, Type, Union, cast, overload
 
 from cognite.client._api_client import APIClient
 from cognite.client._constants import INSTANCES_LIST_LIMIT_DEFAULT
 from cognite.client.data_classes._base import CogniteResourceList
+from cognite.client.data_classes.data_modeling.aggregations import Aggregation, Histogram, HistogramValue
 from cognite.client.data_classes.data_modeling.filters import Filter
 from cognite.client.data_classes.data_modeling.ids import (
     EdgeId,
@@ -20,6 +21,7 @@ from cognite.client.data_classes.data_modeling.instances import (
     EdgeApplyResult,
     EdgeApplyResultList,
     EdgeList,
+    InstanceAggregationResultList,
     InstancesApplyResult,
     InstancesDeleteResult,
     InstanceSort,
@@ -345,7 +347,7 @@ class InstancesAPI(APIClient):
         cls,
         *,
         include_typing: bool,
-        sort: list[InstanceSort | dict] | InstanceSort | dict | None,
+        sort: Sequence[InstanceSort | dict] | InstanceSort | dict | None,
         sources: ViewIdentifier | Sequence[ViewIdentifier] | View | Sequence[View] | None,
         instance_type: Literal["node", "edge"] | None,
     ) -> dict[str, Any]:
@@ -486,13 +488,254 @@ class InstancesAPI(APIClient):
         )
 
     @overload
+    def search(
+        self,
+        view: ViewId,
+        query: str,
+        instance_type: Literal["node"] = "node",
+        properties: list[str] | None = None,
+        filter: Filter | dict | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> NodeList:
+        ...
+
+    @overload
+    def search(
+        self,
+        view: ViewId,
+        query: str,
+        instance_type: Literal["edge"],
+        properties: list[str] | None = None,
+        filter: Filter | dict | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> EdgeList:
+        ...
+
+    def search(
+        self,
+        view: ViewId,
+        query: str,
+        instance_type: Literal["node", "edge"] = "node",
+        properties: list[str] | None = None,
+        filter: Filter | dict | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> NodeList | EdgeList:
+        """`Search instances <https://developer.cognite.com/api/v1/#tag/Instances/operation/searchInstances>`_
+
+        Args:
+            view (ViewId): View to search in.
+            query (str): Query string that will be parsed and used for search.
+            instance_type (Literal["node", "edge"]): Whether to search for nodes or edges.
+            properties (list[str]): Optional array of properties you want to search through.
+                                    If you do not specify one or more properties, the service will
+                                    search all text fields within the view.
+            filter (dict | Filter): Advnanced filtering of instances.
+            limit (int, optional): Maximum number of instances to return. Default to 1000. Set to -1, float("inf") or None
+                to return all items.
+
+        Returns:
+            EdgeList | NodeList: Search result with matching nodes or edges.
+
+        Examples:
+
+            Search for Arnold in the person view in the name property:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling import ViewId
+                >>> c = CogniteClient()
+                >>> res = c.data_modeling.instances.search(ViewId("mySpace", "PersonView", "v1"), query="Arnold", properties=["name"])
+
+            Search for Quentin in the person view in the name property, but only born before 1970:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling import ViewId
+                >>> import cognite.client.data_classes.data_modeling.filters as filters
+                >>> c = CogniteClient()
+                >>> born_after_1970 = filters.Range(["mySpace", "PersonView/v1", "birthYear"], gt=1970)
+                >>> res = c.data_modeling.instances.search(ViewId("mySpace", "PersonView", "v1"),
+                ... query="Quentin", properties=["name"], filter=born_after_1970)
+
+        """
+        if instance_type == "node":
+            list_cls: Union[Type[NodeList], Type[EdgeList]] = NodeList
+        elif instance_type == "edge":
+            list_cls = EdgeList
+        else:
+            raise ValueError(f"Invalid instance type: {instance_type}")
+
+        body = {"view": view.dump(camel_case=True), "query": query, "instanceType": instance_type, "limit": limit}
+        if properties:
+            body["properties"] = properties
+        if filter:
+            body["filter"] = filter.dump() if isinstance(filter, Filter) else filter
+
+        res = self._post(url_path=self._RESOURCE_PATH + "/search", json=body)
+        return list_cls._load(res.json()["items"], cognite_client=None)
+
+    def aggregate(
+        self,
+        view: ViewId,
+        aggregates: Aggregation | dict | Sequence[Aggregation | dict],
+        instance_type: Literal["node", "edge"] = "node",
+        group_by: Sequence[str] | None = None,
+        query: str | None = None,
+        properties: Sequence[str] | None = None,
+        filter: Filter | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> InstanceAggregationResultList:
+        """`Aggregate data across nodes/edges <https://developer.cognite.com/api/v1/#tag/Instances/operation/aggregateInstances>`_
+
+        Args:
+            view (ViewId): View to to aggregate over.
+            aggregates (Aggregation | dict | Sequence[Aggregation | dict]):  The properties to aggregate over.
+            instance_type (Literal["node", "edge"]): Whether to search for nodes or edges.
+            group_by (Optional[Sequence[str]]): The selection of fields to group the results by when doing aggregations.
+                                  You can specify up to 5 items to group by.
+            query (Optional[str]): Query string that will be parsed and used for search.
+            properties (Optional[Sequence[str]]): Optional array of properties you want to search through.
+                                    If you do not specify one or more properties, the service will
+                                    search all text fields within the view.
+            filter (Optional[Filter]): Advnanced filtering of instances.
+            limit (int, optional): Maximum number of instances to return. Default to 1000. Set to -1, float("inf") or None
+                to return all items.
+
+        Returns:
+            InstanceAggregationResultList: Node or edge aggregation results.
+
+        Examples:
+
+            Get the average run time in minutes for movies grouped by release year:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling import ViewId, aggregations as aggs
+                >>> c = CogniteClient()
+                >>> avg_run_time = aggs.Avg("runTimeMinutes")
+                >>> view_id = ViewId("mySpace", "PersonView", "v1")
+                >>> res = c.data_modeling.instances.aggregate(view_id, [avg_run_time], group_by=["releaseYear"])
+
+        """
+        if instance_type not in ("node", "edge"):
+            raise ValueError(f"Invalid instance type: {instance_type}")
+        body: Dict[str, Any] = {"view": view.dump(camel_case=True), "instanceType": instance_type, "limit": limit}
+        aggregate_seq: Sequence[Aggregation | dict] = aggregates if isinstance(aggregates, Sequence) else [aggregates]
+        body["aggregates"] = [
+            agg.dump(camel_case=True) if isinstance(agg, Aggregation) else agg for agg in aggregate_seq
+        ]
+        if group_by:
+            body["groupBy"] = group_by
+        if filter:
+            body["filter"] = filter.dump() if isinstance(filter, Filter) else filter
+        if query:
+            body["query"] = query
+        if properties:
+            body["properties"] = properties
+
+        res = self._post(url_path=self._RESOURCE_PATH + "/aggregate", json=body)
+        return InstanceAggregationResultList._load(res.json()["items"], cognite_client=None)
+
+    @overload
+    def histogram(
+        self,
+        view: ViewId,
+        histograms: Histogram,
+        instance_type: Literal["node", "edge"] = "node",
+        query: str | None = None,
+        properties: Sequence[str] | None = None,
+        filter: Filter | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> HistogramValue:
+        ...
+
+    @overload
+    def histogram(
+        self,
+        view: ViewId,
+        histograms: Sequence[Histogram],
+        instance_type: Literal["node", "edge"] = "node",
+        query: str | None = None,
+        properties: Sequence[str] | None = None,
+        filter: Filter | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> list[HistogramValue]:
+        ...
+
+    def histogram(
+        self,
+        view: ViewId,
+        histograms: Histogram | Sequence[Histogram],
+        instance_type: Literal["node", "edge"] = "node",
+        query: str | None = None,
+        properties: Sequence[str] | None = None,
+        filter: Filter | None = None,
+        limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
+    ) -> HistogramValue | list[HistogramValue]:
+        """`Produces histograms for nodes/edges <https://developer.cognite.com/api/v1/#tag/Instances/operation/aggregateInstances>`_
+
+        Args:
+            view (ViewId): View to to aggregate over.
+            histograms (Histogram | Sequence[Histogram]):  The properties to aggregate over.
+            instance_type (Literal["node", "edge"]): Whether to search for nodes or edges.
+            query (Optional[str]): Query string that will be parsed and used for search.
+            properties (Optional[Sequence[str]]): Optional array of properties you want to search through.
+                                    If you do not specify one or more properties, the service will
+                                    search all text fields within the view.
+            filter (Optional[Filter]): Advnanced filtering of instances.
+            limit (int, optional): Maximum number of instances to return. Default to 1000. Set to -1, float("inf") or None
+                to return all items.
+
+        Returns:
+            list[HistogramValue]: Node or edge aggregation results.
+
+        Examples:
+
+            Find the number of people born per decade:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling import aggregations as aggs, ViewId
+                >>> c = CogniteClient()
+                >>> birth_by_decade = aggs.Histogram("birthYear", interval=10.0)
+                >>> view_id = ViewId("mySpace", "PersonView", "v1")
+                >>> res = c.data_modeling.instances.histogram(view_id, birth_by_decade)
+        """
+        if instance_type not in ("node", "edge"):
+            raise ValueError(f"Invalid instance type: {instance_type}")
+        body: Dict[str, Any] = {"view": view.dump(camel_case=True), "instanceType": instance_type, "limit": limit}
+
+        if isinstance(histograms, Sequence):
+            histogram_seq: Sequence[Histogram] = histograms
+            is_singleton = False
+        elif isinstance(histograms, Histogram):
+            histogram_seq = [histograms]
+            is_singleton = True
+        else:
+            raise TypeError(f"Expected Histogram or sequence of Histograms, got {type(histograms)}")
+
+        for histogram in histogram_seq:
+            if not isinstance(histogram, Histogram):
+                raise ValueError(f"Not a histogram: {histogram}")
+
+        body["aggregates"] = [histogram.dump(camel_case=True) for histogram in histogram_seq]
+        if filter:
+            body["filter"] = filter.dump() if isinstance(filter, Filter) else filter
+        if query:
+            body["query"] = query
+        if properties:
+            body["properties"] = properties
+
+        res = self._post(url_path=self._RESOURCE_PATH + "/aggregate", json=body)
+        if is_singleton:
+            return HistogramValue.load(res.json()["items"][0]["aggregates"][0])
+        else:
+            return [HistogramValue.load(item["aggregates"][0]) for item in res.json()["items"]]
+
+    @overload
     def list(
         self,
         instance_type: Literal["node"] = "node",
         include_typing: bool = False,
         sources: ViewIdentifier | Sequence[ViewIdentifier] | View | Sequence[View] | None = None,
         limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
-        sort: list[InstanceSort | dict] | InstanceSort | dict | None = None,
+        sort: Sequence[InstanceSort | dict] | InstanceSort | dict | None = None,
         filter: Filter | dict | None = None,
     ) -> NodeList:
         ...
@@ -504,7 +747,7 @@ class InstancesAPI(APIClient):
         include_typing: bool = False,
         sources: ViewIdentifier | Sequence[ViewIdentifier] | View | Sequence[View] | None = None,
         limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
-        sort: list[InstanceSort | dict] | InstanceSort | dict | None = None,
+        sort: Sequence[InstanceSort | dict] | InstanceSort | dict | None = None,
         filter: Filter | dict | None = None,
     ) -> EdgeList:
         ...
@@ -515,7 +758,7 @@ class InstancesAPI(APIClient):
         include_typing: bool = False,
         sources: ViewIdentifier | Sequence[ViewIdentifier] | View | Sequence[View] | None = None,
         limit: int = INSTANCES_LIST_LIMIT_DEFAULT,
-        sort: list[InstanceSort | dict] | InstanceSort | dict | None = None,
+        sort: Sequence[InstanceSort | dict] | InstanceSort | dict | None = None,
         filter: Filter | dict | None = None,
     ) -> NodeList | EdgeList:
         """`List instances <https://developer.cognite.com/api#tag/Instances/tag/Instances/operation/advancedListInstance>`_
@@ -526,7 +769,7 @@ class InstancesAPI(APIClient):
             sources (ViewIdentifier | Sequence[ViewIdentifier] | View | Sequence(View) | None): Views to retrieve properties from.
             limit (int, optional): Maximum number of instances to return. Default to 1000. Set to -1, float("inf") or None
                 to return all items.
-            sort (list[InstanceSost] | InstanceSort | dict): How you want the listed instances information ordered.
+            sort (Sequence[InstanceSost] | InstanceSort | dict): How you want the listed instances information ordered.
             filter (dict | Filter): Advanced filtering of instances.
 
         Returns: