+++ cognite/client/data_classes/data_modeling/instances.py
@@ -30,6 +30,7 @@ from cognite.client.data_classes._base import (
 )
 from cognite.client.data_classes.data_modeling._core import DataModelingResource
 from cognite.client.data_classes.data_modeling._validation import validate_data_modeling_identifier
+from cognite.client.data_classes.data_modeling.aggregations import AggregatedNumberedValue
 from cognite.client.data_classes.data_modeling.data_types import (
     DirectRelationReference,
 )
@@ -303,6 +304,59 @@ class InstanceApplyResult(InstanceCore):
         self.was_modified = was_modified
         self.last_updated_time = last_updated_time
         self.created_time = created_time
+
+
+class InstanceAggregationResult(DataModelingResource):
+    """A node or edge. This represents the update on the instance.
+
+    Args:
+        aggregates (list[AggregatedNumberedValue]) : List of aggregated values.
+        group (dict[str, str | int | float | bool]) : The grouping used for the aggregation.
+    """
+
+    def __init__(self, aggregates: list[AggregatedNumberedValue], group: dict[str, str | int | float | bool]):
+        self.aggregates = aggregates
+        self.group = group
+
+    @classmethod
+    def load(cls, data: dict | str) -> InstanceAggregationResult:
+        """
+        Loads an instance from a json string or dictionary.
+
+        Args:
+            data (dict | str): The json string or dictionary.
+
+        Returns:
+            An instance.
+
+        """
+        data = json.loads(data) if isinstance(data, str) else data
+
+        return cls(
+            aggregates=[AggregatedNumberedValue.load(agg) for agg in data["aggregates"]],
+            group=cast(Dict[str, Union[str, int, float, bool]], data.get("group")),
+        )
+
+    def dump(self, camel_case: bool = False) -> dict[str, Any]:
+        """
+        Dumps the instance to a dictionary.
+
+        Args:
+            camel_case (bool): Whether to convert the keys to camel case.
+
+        Returns:
+            A dictionary.
+
+        """
+        return {
+            "aggregates": [agg.dump(camel_case) for agg in self.aggregates],
+            "group": self.group,
+            ("instanceType" if camel_case else "instance_type"): self.instance_type,
+        }
+
+
+class InstanceAggregationResultList(CogniteResourceList[InstanceAggregationResult]):
+    _RESOURCE = InstanceAggregationResult
 
 
 class NodeApply(InstanceApply):