+++ tests/tests_integration/test_api/test_data_modeling/test_instances.py
@@ -20,8 +20,10 @@ from cognite.client.data_classes.data_modeling import (
     SingleHopConnectionDefinition,
     Space,
     View,
+    aggregations,
     filters,
 )
+from cognite.client.data_classes.data_modeling.aggregations import HistogramValue
 from cognite.client.data_classes.data_modeling.filters import Equals
 from cognite.client.exceptions import CogniteAPIError
 
@@ -52,6 +54,11 @@ def person_view(cognite_client: CogniteClient, integration_test_space: Space) ->
 @pytest.fixture()
 def actor_view(cognite_client: CogniteClient, integration_test_space: Space) -> View:
     return cognite_client.data_modeling.views.retrieve((integration_test_space.space, "Actor", "2"))[0]
+
+
+@pytest.fixture()
+def movie_view(cognite_client: CogniteClient, integration_test_space: Space) -> View:
+    return cognite_client.data_modeling.views.retrieve((integration_test_space.space, "Movie", "2"))[0]
 
 
 class TestInstancesAPI:
@@ -349,3 +356,63 @@ class TestInstancesAPI:
         finally:
             # Cleanup
             cognite_client.data_modeling.instances.delete(valid_person.as_id())
+
+    def test_search_node_data(self, cognite_client: CogniteClient, person_view: View) -> None:
+        # Act
+        search_result = cognite_client.data_modeling.instances.search(
+            person_view.as_id(), query="Quentin", properties=["name"]
+        )
+
+        # Assert
+        assert len(search_result) == 1
+        assert search_result[0].external_id == "person:quentin_tarantino"
+
+    def test_search_node_data_with_invalid_property(self, cognite_client: CogniteClient, person_view: View) -> None:
+        # Act
+        with pytest.raises(CogniteAPIError) as error:
+            cognite_client.data_modeling.instances.search(
+                person_view.as_id(), query="Quentin", properties=["invalidProperty"]
+            )
+
+        # Assert
+        assert "Unknown property" in error.value.message
+
+    def test_search_node_data_with_filtering(self, cognite_client: CogniteClient, person_view: View) -> None:
+        # Act
+        view_id = person_view.as_id()
+        f = filters
+        born_after_2000 = f.Range([view_id.space, view_id.as_source_identifier(), "birthYear"], gt=2000)
+
+        # Act
+        search_result = cognite_client.data_modeling.instances.search(
+            view_id, query="Quentin", properties=["name"], filter=born_after_2000
+        )
+
+        # Assert
+        assert len(search_result) == 0
+
+    def test_aggregate_histogram_across_nodes(self, cognite_client: CogniteClient, person_view: View) -> None:
+        view_id = person_view.as_id()
+        birth_by_decade = aggregations.Histogram("birthYear", interval=10.0)
+
+        histogram = cognite_client.data_modeling.instances.histogram(view_id, birth_by_decade)
+        assert isinstance(histogram, HistogramValue)
+
+        histogram_seq = cognite_client.data_modeling.instances.histogram(view_id, [birth_by_decade])
+        assert len(histogram_seq) == 1 and isinstance(histogram_seq[0], HistogramValue)
+
+    def test_aggregate_with_grouping(self, cognite_client: CogniteClient, movie_view: View) -> None:
+        # Arrange
+        view_id = movie_view.as_id()
+        avg_agg = aggregations.Avg("runTimeMinutes")
+        max_agg = aggregations.Max("runTimeMinutes")
+
+        # Act
+        counts = cognite_client.data_modeling.instances.aggregate(
+            view_id,
+            aggregates=[avg_agg, max_agg],
+            group_by=["releaseYear"],
+        )
+
+        # Assert
+        assert len(counts)